release Linux

+++ client/include/Definitions.hh
@@ -34,8 +34,8 @@ namespace client
     };
 }
 
-typedef std::pair<int16_t, int16_t>	pos_t;
-typedef std::pair<int16_t, int16_t>	vec_t;
+typedef std::pair<int32_t, int32_t>	pos_t;
+typedef std::pair<int32_t, int32_t>	vec_t;
 typedef uint32_t			tick;
 typedef uint32_t			ide_t;
 typedef uint16_t			typeide_t;
