+++ client/include/NetworkManager.hh
@@ -46,8 +46,8 @@ namespace client {
 
       void receivePlayerData(uint32_t playerId, uint8_t nbAttack);
 
-        void receiveMoveEntity(uint32_t tick, uint32_t eventId, uint32_t entityId, int16_t vecx, int16_t vecy,
-                               int16_t posx, int16_t posy);
+        void receiveMoveEntity(uint32_t tick, uint32_t eventId, uint32_t entityId, int32_t vecx, int32_t vecy,
+                               int32_t posx, int32_t posy);
 
         void receivePlaySound(uint32_t tick, uint32_t eventId, const std::string &SoundName);
 
@@ -55,8 +55,8 @@ namespace client {
 
         void
         receiveSpawnEntity(uint32_t tick, uint32_t eventId, const std::string &spriteName, uint32_t entityId,
-                           int16_t pos_x,
-                           int16_t pos_y, int16_t hp);
+                           int32_t pos_x,
+                           int32_t pos_y, int16_t hp);
 
         void receiveUpdateEntity(uint32_t tick, uint32_t eventId, uint32_t entityId, int16_t hp);
 
@@ -74,7 +74,7 @@ namespace client {
 
         void sendQuit();
         
-        void sendPlayerMove(uint32_t tick, int16_t vect_x, int16_t vect_y, int16_t pos_x, int16_t pos_y);
+        void sendPlayerMove(uint32_t tick, int32_t vect_x, int32_t vect_y, int32_t pos_x, int32_t pos_y);
 
         void sendPlayerAttack(int32_t tick, uint8_t attackId);
