+++ core/include/network/packet/PacketMoveEntity.hh
@@ -13,21 +13,21 @@ namespace network {
         class PacketMoveEntity : public APacket {
         public:
             virtual ~PacketMoveEntity();
-            PacketMoveEntity(uint32_t tick = 0,uint32_t eventId = 0, uint32_t entityId = 0, int16_t vecx = 0, int16_t vecy = 0, int16_t posx = 0, int16_t posy = 0);
+            PacketMoveEntity(uint32_t tick = 0,uint32_t eventId = 0, uint32_t entityId = 0, int32_t vecx = 0, int32_t vecy = 0, int32_t posx = 0, int32_t posy = 0);
             void setTick(uint32_t tick);
             uint32_t getTick() const;
             void setEventId(uint32_t id);
             uint32_t getEventId() const;
             void setEntityId(uint32_t entityId);
             uint32_t getEntityId() const;
-            void setVecX(int16_t vecX);
-            int16_t getVecX() const;
-            void setVecY(int16_t vecY);
-            int16_t getVecY() const;
-            void setPosX(int16_t pos_x);
-            int16_t  getPosX() const;
-            void setPosY(int16_t pos_y);
-            int16_t getPosY() const;
+            void setVecX(int32_t vecX);
+            int32_t getVecX() const;
+            void setVecY(int32_t vecY);
+            int32_t getVecY() const;
+            void setPosX(int32_t pos_x);
+            int32_t  getPosX() const;
+            void setPosY(int32_t pos_y);
+            int32_t getPosY() const;
         };
     }
 }
